orchestrator/requirements.txt
python-socketio
requests
pyaudio
websockets
python-dotenv
openwakeword
numpy
scipy
